requirements.txt
streamlit==1.36.0
langchain==0.2.0
langchain-google-genai==0.1.0
google-generativeai==0.3.2
PyPDF2==3.0.1
python-dotenv==1.0.0
langchain-community==0.2.0
faiss-cpu==1.7.4
sentence-transformers==2.7.0
huggingface-hub==0.23.0